Add files via upload
Language : C++

Title : Luhn’s Algorithm

Description:
This is small program that checks  if the given number is valid or not.

Rules followed:
(1). Multiply every other digit by 2, starting with the number’s second-to-last digit, and then add those products’ digits together.
(2). Add the sum to the sum of the digits that weren’t multiplied by 2.
(3). If the total’s last digit is 0 (or, put more formally, if the total modulo 10 is congruent to 0), the number is valid!

Restrictions:
Only checks for the following card:
     (1). VISA
     (2). MASTERCARD (Only the ones starting with 51, 52, 53, 54, 55)
     (3). AMERICAN EXPRESS (AMEX)